Move pretrained config
