+++ lerobot/common/policies/act/configuration_act.py
@@ -16,8 +16,9 @@
 from dataclasses import dataclass, field
 
 from lerobot.common.optim.optimizers import AdamWConfig
-from lerobot.configs.policies import PreTrainedConfig
 from lerobot.configs.types import NormalizationMode
+
+from ..config import PreTrainedConfig
 
 
 @PreTrainedConfig.register_subclass("act")
