+++ lerobot/common/policies/diffusion/configuration_diffusion.py
@@ -18,8 +18,9 @@ from dataclasses import dataclass, field
 
 from lerobot.common.optim.optimizers import AdamConfig
 from lerobot.common.optim.schedulers import DiffuserSchedulerConfig
-from lerobot.configs.policies import PreTrainedConfig
 from lerobot.configs.types import NormalizationMode
+
+from ..config import PreTrainedConfig
 
 
 @PreTrainedConfig.register_subclass("diffusion")
