+++ lerobot/common/policies/pretrained.py
@@ -26,8 +26,8 @@ from safetensors.torch import load_model as load_model_as_safetensor
 from safetensors.torch import save_model as save_model_as_safetensor
 from torch import Tensor, nn
 
+from lerobot.common.policies.config import PreTrainedConfig
 from lerobot.common.utils.hub import HubMixin
-from lerobot.configs.policies import PreTrainedConfig
 
 T = TypeVar("T", bound="PreTrainedPolicy")
