+++ lerobot/common/policies/tdmpc/configuration_tdmpc.py
@@ -17,8 +17,9 @@
 from dataclasses import dataclass, field
 
 from lerobot.common.optim.optimizers import AdamConfig
-from lerobot.configs.policies import PreTrainedConfig
 from lerobot.configs.types import NormalizationMode
+
+from ..config import PreTrainedConfig
 
 
 @PreTrainedConfig.register_subclass("tdmpc")
