+++ lerobot/common/policies/vqbet/configuration_vqbet.py
@@ -20,8 +20,9 @@ from dataclasses import dataclass, field
 
 from lerobot.common.optim.optimizers import AdamConfig
 from lerobot.common.optim.schedulers import VQBeTSchedulerConfig
-from lerobot.configs.policies import PreTrainedConfig
 from lerobot.configs.types import NormalizationMode
+
+from ..config import PreTrainedConfig
 
 
 @PreTrainedConfig.register_subclass("vqbet")
