+++ lerobot/common/robot_devices/control_configs.py
@@ -17,9 +17,9 @@ from pathlib import Path
 
 import draccus
 
+from lerobot.common.policies.config import PreTrainedConfig
 from lerobot.common.robot_devices.robots.configs import RobotConfig
 from lerobot.configs import parser
-from lerobot.configs.policies import PreTrainedConfig
 
 
 @dataclass
