+++ tests/test_control_robot.py
@@ -42,6 +42,7 @@ from unittest.mock import patch
 import pytest
 
 from lerobot.common.policies.act.configuration_act import ACTConfig
+from lerobot.common.policies.config import PreTrainedConfig
 from lerobot.common.policies.factory import make_policy
 from lerobot.common.robot_devices.control_configs import (
     CalibrateControlConfig,
@@ -49,7 +50,6 @@ from lerobot.common.robot_devices.control_configs import (
     ReplayControlConfig,
     TeleoperateControlConfig,
 )
-from lerobot.configs.policies import PreTrainedConfig
 from lerobot.scripts.control_robot import calibrate, record, replay, teleoperate
 from tests.test_robots import make_robot
 from tests.utils import TEST_ROBOT_TYPES, mock_calibration_dir, require_robot
